Passing by Reference
To pass a value by reference, argument pointers are passed to the functions just like any other value
You need to declare the function parameters as pointer types. Changes within the function are reflected outside the function as well. 
When we assign inside this function, we must use the indirection operator to dereference cause we want to assign the values.